refactor: Relayer approval composable
